+++ CHANGES.md
@@ -11,6 +11,7 @@ Change Log
 * `mapItems` can now return arbitrary Cesium primitives.
 * Update `terriajs-cesium` to `1.91.0`.
 * Update `terriajs-cesium` to `1.92.0`.
+* Added support for using Compare / Split Screen mode with Cesium 3D Tiles.
 * [The next improvement]
 
 #### 8.2.4 - 2022-05-23